+++ src/server/resc_def_all.c
@@ -754,9 +754,9 @@ int init_resc_defs(void)
 #ifndef PBS_MOM
   /* build up a temporary list of string resources */
 
-  if (server.sv_attr[(int)SRV_ATR_ExtraResc].at_flags & ATR_VFLAG_SET)
+  if (server.sv_attr[SRV_ATR_ExtraResc].at_flags & ATR_VFLAG_SET)
     {
-    resc_arst = server.sv_attr[(int)SRV_ATR_ExtraResc].at_val.at_arst;
+    resc_arst = server.sv_attr[SRV_ATR_ExtraResc].at_val.at_arst;
 
     tmpresc = calloc(resc_arst->as_usedptr + 1, sizeof(resource_def));
